Build: add make rpms target and convenience script makerpms.sh
make rpms and ./makerpms.sh will produce the same RPM packages. The
advantage of makerpms.sh is that it will take care of initial
autoreconf & configure phases as needed.

There seems to be an issue either in GNU Make or the way rpmbuild is
calling it. make rpms fails if MAKEFLAGS environment variable specifies
parallel build. The error message is like:
  INTERNAL: Exiting with 1 jobserver tokens available; should be 8!

https://fedorahosted.org/freeipa/ticket/6418